pyproject.toml
[project]
name = "PolyKalshiTradingInterface"                 # Name of your top-level package (i.e. folder name)
version = "0.1.0"
description = "Please read the README and SETUP"
authors = [
  { name = "Rohit Dayanand", email = "rohitdayanand8@gmail.com" }
]
dependencies = [
  "asyncpg==0.30.0",
  "certifi==2025.4.26",
  "cffi==1.17.1",
  "charset-normalizer==3.4.2",
  "DateTime==5.5",
  "gevent==25.5.1",
  "greenlet==3.2.2",
  "idna==3.10",
  "numpy==2.2.6",
  "pandas==2.2.3",
  "pyasn1==0.6.1",
  "pycparser==2.22",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.0.1",
  "pytz==2025.2",
  "requests==2.32.3",
  "rsa==4.9.1",
  "setuptools==80.7.1",
  "six==1.17.0",
  "SQLAlchemy==2.0.41",
  "cryptography",
  "psycopg[async]",
  "websockets==15.0.1",
  "websocket-client",
  "pyee==12.1.1",
  "fastapi==0.115.6",
  "uvicorn==0.34.0"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
